Extract TUI MCP startup handling
