+++ codex-rs/tui/src/chatwidget/tests.rs
@@ -170,9 +170,6 @@ pub(super) use codex_protocol::protocol::GuardianRiskLevel;
 pub(super) use codex_protocol::protocol::GuardianUserAuthorization;
 pub(super) use codex_protocol::protocol::ImageGenerationEndEvent;
 pub(super) use codex_protocol::protocol::ItemCompletedEvent;
-pub(super) use codex_protocol::protocol::McpStartupCompleteEvent;
-pub(super) use codex_protocol::protocol::McpStartupStatus;
-pub(super) use codex_protocol::protocol::McpStartupUpdateEvent;
 pub(super) use codex_protocol::protocol::ModelVerification as CoreModelVerification;
 pub(super) use codex_protocol::protocol::ModelVerificationEvent;
 pub(super) use codex_protocol::protocol::NonSteerableTurnKind;
